requirements.txt
flask==3.1.0
gunicorn==23.0.0
python-telegram-bot==21.10
requests==2.32.3
python-dotenv==1.1.0
anthropic==0.52.0
pytest==8.3.4
